pyproject.toml
[project]
name = "benfica-discord-bot"
version = "0.1.0"
description = "Discord bot for posting Portuguese sports newspaper covers and Benfica match information"
authors = [
    {name = "Edgar Gonçalves", email = "edgar.g.goncalves@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.1.0",
    "beautifulsoup4>=4.11.0",
    "requests>=2.28.2",
    "apscheduler>=3.9.1",
    "charset-normalizer>=2.1.1",
    "aiohttp>=3.8.3",
    "selenium>=4.7.2",
    "webdriver-manager>=3.8.5",
    "pendulum>=2.1.2",
    "pillow>=10.0.1",
    "fake-useragent>=1.4.0",
    "python-dotenv>=1.0.0",
    "curl-cffi>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear (common bugs)
    "S",   # flake8-bandit (security issues)
    "C90", # mccabe (complexity)
]

# Ignore specific rules that are too strict for this project
ignore = [
    "S101",  # Allow assert statements (used in tests)
    "S105",  # Allow hardcoded passwords in tests
    "S108",  # Allow /tmp usage (secure enough for our use case)
    "S603",  # Allow subprocess without shell=True validation
    "S607",  # Allow starting processes with partial paths
]

[tool.ruff.lint.mccabe]
max-complexity = 10  # Maximum cyclomatic complexity

[tool.ruff.lint.per-file-ignores]
# ESPN scraper has complex parsing logic that's acceptable
"core/match/sources.py" = ["C901"]
# Test files can have long lines in HTML fixtures
"tests/test_core/test_benfica_calendar.py" = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]
